backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
celery==5.3.6
redis==5.0.1
anthropic==0.18.0
openai==1.12.0
PyGithub==2.2.0
python-telegram-bot==20.8
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
python-multipart==0.0.9
pdfplumber==0.10.3
python-docx==1.1.0
nltk==3.8.1
boto3==1.34.0
pgvector==0.2.4
cryptography==42.0.0
